backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
